--- a/src/decode/mod.rs
+++ b/src/decode/mod.rs
@@ -5,5 +5,6 @@ pub mod lzma;
 pub mod lzma2;
 pub mod options;
 pub mod rangecoder;
+pub mod stream;
 pub mod util;
 pub mod xz;