--- a/src/decode/rangecoder.rs
+++ b/src/decode/rangecoder.rs
@@ -28,9 +28,51 @@ where
         Ok(dec)
     }
 
+    pub fn from_parts(stream: &'a mut R, range: u32, code: u32) -> Self {
+        Self {
+            stream,
+            range,
+            code,
+        }
+    }
+
+    pub fn into_parts(self) -> (u32, u32) {
+        (self.range, self.code)
+    }
+
+    pub fn set(&mut self, range: u32, code: u32) {
+        self.range = range;
+        self.code = code;
+    }
+
+    pub fn range(&self) -> u32 {
+        self.range
+    }
+
+    pub fn code(&self) -> u32 {
+        self.code
+    }
+
+    pub fn buf(&mut self) -> io::Result<&[u8]> {
+        self.stream.fill_buf()
+    }
+
+    pub fn remaining(&mut self) -> io::Result<usize> {
+        Ok(self.buf()?.len())
+    }
+
+    pub fn read_into(&mut self, dst: &mut [u8]) -> io::Result<usize> {
+        self.stream.read(dst)
+    }
+
     #[inline]
     pub fn is_finished_ok(&mut self) -> io::Result<bool> {
-        Ok(self.code == 0 && util::is_eof(self.stream)?)
+        Ok(self.code == 0 && self.is_eof()?)
+    }
+
+    #[inline]
+    pub fn is_eof(&mut self) -> io::Result<bool> {
+        util::is_eof(self.stream)
     }
 
     #[inline]
@@ -92,10 +134,36 @@ where
         }
     }
 
+    #[inline]
+    pub fn decode_bit_check(&mut self, prob: u16) -> io::Result<bool> {
+        let bound: u32 = (self.range >> 11) * (prob as u32);
+        if self.code < bound {
+            self.range = bound;
+
+            self.normalize()?;
+            Ok(false)
+        } else {
+            self.code -= bound;
+            self.range -= bound;
+
+            self.normalize()?;
+            Ok(true)
+        }
+    }
+
     fn parse_bit_tree(&mut self, num_bits: usize, probs: &mut [u16]) -> io::Result<u32> {
         let mut tmp: u32 = 1;
         for _ in 0..num_bits {
             let bit = self.decode_bit(&mut probs[tmp as usize])?;
+            tmp = (tmp << 1) ^ (bit as u32);
+        }
+        Ok(tmp - (1 << num_bits))
+    }
+
+    fn parse_bit_tree_check(&mut self, num_bits: usize, probs: &[u16]) -> io::Result<u32> {
+        let mut tmp: u32 = 1;
+        for _ in 0..num_bits {
+            let bit = self.decode_bit_check(probs[tmp as usize])?;
             tmp = (tmp << 1) ^ (bit as u32);
         }
         Ok(tmp - (1 << num_bits))
@@ -111,6 +179,22 @@ where
         let mut tmp: usize = 1;
         for i in 0..num_bits {
             let bit = self.decode_bit(&mut probs[offset + tmp])?;
+            tmp = (tmp << 1) ^ (bit as usize);
+            result ^= (bit as u32) << i;
+        }
+        Ok(result)
+    }
+
+    pub fn parse_reverse_bit_tree_check(
+        &mut self,
+        num_bits: usize,
+        probs: &[u16],
+        offset: usize,
+    ) -> io::Result<u32> {
+        let mut result = 0u32;
+        let mut tmp: usize = 1;
+        for i in 0..num_bits {
+            let bit = self.decode_bit_check(probs[offset + tmp])?;
             tmp = (tmp << 1) ^ (bit as usize);
             result ^= (bit as u32) << i;
         }
@@ -137,11 +221,22 @@ impl BitTree {
         rangecoder.parse_bit_tree(self.num_bits, self.probs.as_mut_slice())
     }
 
+    pub fn parse_check<R: io::BufRead>(&self, rangecoder: &mut RangeDecoder<R>) -> io::Result<u32> {
+        rangecoder.parse_bit_tree_check(self.num_bits, self.probs.as_slice())
+    }
+
     pub fn parse_reverse<R: io::BufRead>(
         &mut self,
         rangecoder: &mut RangeDecoder<R>,
     ) -> io::Result<u32> {
         rangecoder.parse_reverse_bit_tree(self.num_bits, self.probs.as_mut_slice(), 0)
+    }
+
+    pub fn parse_reverse_check<R: io::BufRead>(
+        &self,
+        rangecoder: &mut RangeDecoder<R>,
+    ) -> io::Result<u32> {
+        rangecoder.parse_reverse_bit_tree_check(self.num_bits, self.probs.as_slice(), 0)
     }
 }
 
@@ -175,6 +270,20 @@ impl LenDecoder {
             Ok(self.mid_coder[pos_state].parse(rangecoder)? as usize + 8)
         } else {
             Ok(self.high_coder.parse(rangecoder)? as usize + 16)
+        }
+    }
+
+    pub fn decode_check<R: io::BufRead>(
+        &self,
+        rangecoder: &mut RangeDecoder<R>,
+        pos_state: usize,
+    ) -> io::Result<usize> {
+        if !rangecoder.decode_bit_check(self.choice)? {
+            Ok(self.low_coder[pos_state].parse_check(rangecoder)? as usize)
+        } else if !rangecoder.decode_bit_check(self.choice2)? {
+            Ok(self.mid_coder[pos_state].parse_check(rangecoder)? as usize + 8)
+        } else {
+            Ok(self.high_coder.parse_check(rangecoder)? as usize + 16)
         }
     }
 }